feat(account): migrate base to ts
